Remove serve pre-compressed CSS/JS rules (PR #2779)
* update version to 3.6.0.3
* deprecate compressed rules
* delete related tests
