+++ inc/functions/htaccess.php
@@ -142,7 +142,6 @@ function get_rocket_htaccess_marker() { // phpcs:ignore WordPress.NamingConventi
 	$marker .= get_rocket_htaccess_web_fonts_access();
 	$marker .= get_rocket_htaccess_files_match();
 	$marker .= get_rocket_htaccess_mod_expires();
-	$marker .= rocket_get_compressed_assets_rules();
 	$marker .= get_rocket_htaccess_mod_deflate();
 
 	if ( \WP_Rocket\Buffer\Cache::can_generate_caching_files() && ! is_rocket_generate_caching_mobile_files() ) {
@@ -639,54 +638,6 @@ function get_rocket_htaccess_web_fonts_access() { // phpcs:ignore WordPress.Nami
 	$rules = apply_filters( 'rocket_htaccess_web_fonts_access', $rules );
 
 	return $rules;
-}
-
-/**
- * Rules to serve gzip compressed CSS & JS files if they exists and client accepts gzip
- *
- * @since 3.6.0.2 Update rules used to prevent content encoding issue
- * @since 3.6
- * @author Remy Perona
- *
- * @return string
- */
-function rocket_get_compressed_assets_rules() {
-	$rules = <<<HTACCESS
-<IfModule mod_headers.c>
-    RewriteCond %{HTTP:Accept-Encoding} gzip
-    RewriteCond %{REQUEST_FILENAME}\.gz -f
-    RewriteRule \.(css|js)$ %{REQUEST_URI}.gz [L]
-
-    # Prevent mod_deflate double gzip
-	RewriteRule \.gz$ - [E=no-gzip:1]
-
-	<FilesMatch "\.gz$">
-
-        # Serve correct content types
-        <IfModule mod_mime.c>
-            # (1)
-            RemoveType gz
-
-            # Serve correct content types
-            AddType text/css              css.gz
-            AddType text/javascript       js.gz
-
-            # Serve correct content charset
-            AddCharset utf-8 .css.gz \
-                             .js.gz
-		</IfModule>
-
-        # Force proxies to cache gzipped and non-gzipped files separately
-        Header append Vary Accept-Encoding
-	</FilesMatch>
-
-    # Serve correct encoding type
-    AddEncoding gzip .gz
-</IfModule>
-
-HTACCESS;
-
-	return apply_filters( 'rocket_htaccess_compressed_assets', $rules );
 }
 
 /**
